clean up definitions of ctod(), dtoc(), ctob(), btoc(), btodb(), and dbtob().  remove unused definitions of ctos(), stoc(), and dtob().

+++ sys/arch/hp300/include/param.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: param.h,v 1.20 1995/03/28 18:16:35 jtc Exp $	*/
+/*	$NetBSD: param.h,v 1.21 1995/06/26 06:55:54 cgd Exp $	*/
 
 /*
  * Copyright (c) 1988 University of Utah.
@@ -110,13 +110,13 @@
 #define	NKMEMCLUSTERS	(2048 * 1024 / CLBYTES)
 #endif
 
-/* pages ("clicks") (4096 bytes) to disk blocks */
-#define	ctod(x)		((unsigned)(x) << (PGSHIFT - DEV_BSHIFT))
-#define	dtoc(x)		((unsigned)(x) >> (PGSHIFT - DEV_BSHIFT))
+/* pages ("clicks") to disk blocks */
+#define	ctod(x)		((x) << (PGSHIFT - DEV_BSHIFT))
+#define	dtoc(x)		((x) >> (PGSHIFT - DEV_BSHIFT))
 
 /* pages to bytes */
-#define	ctob(x)		((unsigned)(x) << PGSHIFT)
-#define	btoc(x)		(((unsigned)(x) + PGOFSET) >> PGSHIFT)
+#define	ctob(x)		((x) << PGSHIFT)
+#define	btoc(x)		(((x) + PGOFSET) >> PGSHIFT)
 
 /* bytes to disk blocks */
 #define	dbtob(x)	((x) << DEV_BSHIFT)
